RATIS-473. Clarify size and length methods on LogStream
getLength() was meant to be the number of records in a log
but was implemented as the size in bytes of the records in
the log. This commit adds a getSize() which takes the old
implementation of getLength(), and then reimplements
getLength() as the number of records in the log.

Signed-off-by: Vladimir Rodionov <vrodionov@apache.org>
Signed-off-by: Rajeshbabu Chintaguntla <rajeshbabu@apache.org>
(cherry picked from commit aa6ed83a1940dec05f021c3340b316d3f7bce00a)

Change-Id: I81a927333f18527fdc48aef89e38a3fa85ee655e